added create vector sign hot key
crtl+b hot key creates a vector
and i also fixed the enter to create a new line with the correct information. it was slightly off but now everything works good. I also moved the html text that makes a editor-line into templates. I switched out the onclick event for a mathfield to focus to mouse down and made it inline in the html rather than something that is binding in the javascript. works better because people can select information in a math field without having to click into it first