added a new color scheme that i like
also added message to import variables modal explaining the benefits of importing a variable. Fixed the code that makes equations smaller if they go outside of the bounds of their columns in the list of equations. Deleted nav bar and placed a new skinnier one that takes up less space and looks better. moved the onclick event to open up import variable definition to the name of the area of physics in physics equations