added more specific error handling for trig functions and made sure trig functions work and are rendered properly
added support for trig functions raised to a power so a user can type sin^2() and we will understand that it is (sin())^2
additionally fixed the "ghost variable" issue. where variables would appear in the variables collection twice. I think this was an issue with how debounce which in a way is kinda asynchronous would be editing an object but also reading from it  in another place in the code which would cause wierd things to popup and ultimately break the application