Rethought how vector and vector magnitudes should interact with each other
Added code so that if a vector is defined then if you use its magnitude its units will automatically be defined and vise versa.
Removed the whole VectorMagnitudes system because this system works better.
Change UI so vector and vector magnitudes appear on the same line and you edit their units at the same time and their state and value seperately.
Added some extra data to save to localStorage so that undefined variables that are given and or have a value can be saved and that data can be retrieved once the user refreshes.